+++ azure_terminal/core/system_commands/clear_command.py
@@ -1,0 +1,7 @@
+import os
+
+def execute_clear(prompt):
+    if os.name == 'nt':
+        os.system('cls')
+    else:
+        os.system('clear')