+++ azure_terminal/core/system_commands/quit_command.py
@@ -1,3 +1,3 @@
 def execute_quit(prompt):
-    print(f"Quitting following {prompt}...")
+    print("Quitting...")
     exit()