+++ azure_terminal/module_tests/math_module_test.py
@@ -19,9 +19,10 @@ def math_module_tests():
     while True:
         times_to_run: str = str(input("Enter how many times to run test - "))
         maximum_number: int = int(input("Enter the maximum number for random number generator - "))
+        rounding_off: str = str(input("Enter a number to round off too, input '-' for no rounding - "))
         if times_to_run.isdigit():
             for i in range(0,int(times_to_run)):
-                prompt: str = f"math {random.randint((maximum_number*-1),maximum_number)} {random_operator_generator()} {random.randint((maximum_number*-1),maximum_number)}"
+                prompt: str = f"math {random.randint((maximum_number*-1),maximum_number)} {random_operator_generator()} {random.randint((maximum_number*-1),maximum_number)} {rounding_off}"
                 execute_math(prompt)
             exit()
         else: